/EXEC crashed with 64bit systems. Patch by Soren Jacobsen
git-svn-id: http://svn.irssi.org/repos/irssi/trunk@3272 dbcabf3a-b0e7-0310-adc4-f8d773084564

+++ src/fe-common/core/fe-exec.c
@@ -349,7 +349,7 @@ static void process_exec(PROCESS_REC *rec, const char *cmd)
 static void sig_exec_input_reader(PROCESS_REC *rec)
 {
         char tmpbuf[512], *str;
-        unsigned int recvlen;
+        gsize recvlen;
 	int err, ret;
 
 	g_return_if_fail(rec != NULL);
